kaleo/booking_app/backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
mysqlclient==2.2.0
python-decouple==3.8
gunicorn==21.2.0
django-extensions==4.0
reportlab==4.0.4  # Para generación de PDFs
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0